requirements.txt
astroid==2.3.2
autopep8==1.4.4
beautifulsoup4==4.8.0
Django==2.2.5
django-bootstrap4==1.0.1
django-filter==2.2.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
numpy==1.17.3
pandas==0.25.3
pycodestyle==2.5.0
pylint==2.4.3
python-dateutil==2.8.1
pytz==2019.2
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.5
wrapt==1.11.2
zenhan==0.5.2
